+++ apps/epistola/src/main/resources/db/migration/V2__create_document_templates.sql
@@ -3,7 +3,9 @@ CREATE TABLE document_templates (
     id BIGSERIAL PRIMARY KEY,
     tenant_id BIGINT NOT NULL REFERENCES tenants(id) ON DELETE CASCADE,
     name VARCHAR(255) NOT NULL,
-    content jsonb,
+    template_model jsonb,
+    data_model jsonb,
+    data_examples jsonb DEFAULT '[]'::jsonb,
     created_at TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW(),
     last_modified TIMESTAMP WITH TIME ZONE NOT NULL DEFAULT NOW()
 );